init of subcommands inside corresponding files (#1);
